Added fire, adjusted volume dynamically, and some cleanup
Took out pausing automatically when you leave the window (who knows, maybe somebody wants to listen to the music). Removed some dead code.

Adjusted the build a bit to try to shave off a few more bytes. Definitely going to need to start looking for things to reduce size more, though the DEER and BISON are close to ready so it might not be bad.

56876 bytes raw
26377 bytes minified
23944 bytes tokenized
24202 bytes built
Built zip 10504/13312 bytes (78.9%)